Move concurrent samplers to explorers
